Add filter helpers and integrate with practice plan list
